using fabsf instead of abs
disable ipv6

using old version of openal

fix SYS.timeBase function call

fix download_complete_data_t in sys_common.js

disable git checking for now

disable openal interval update for now

+++ code/client/cl_cin.c
@@ -1357,7 +1357,7 @@ e_status CIN_RunCinematic (int handle)
 	}
 
 	thisTime = CL_ScaledMilliseconds()*com_timescale->value;
-	if (cinTable[currentHandle].shader && (abs(thisTime - cinTable[currentHandle].lastTime))>100) {
+	if (cinTable[currentHandle].shader && (abs(thisTime - (int)(cinTable[currentHandle].lastTime) ))>100) {
 		cinTable[currentHandle].startTime += thisTime - cinTable[currentHandle].lastTime;
 	}
 	cinTable[currentHandle].tfps = ((((CL_ScaledMilliseconds()*com_timescale->value) - cinTable[currentHandle].startTime)*3)/100);
